core: Move EstimateFee out of server and into core.
No functional changes.
