+++ server/segment_rpc.go
@@ -7,7 +7,6 @@ import (
 	"encoding/base64"
 	"fmt"
 	"io/ioutil"
-	"math"
 	"math/big"
 	"net/http"
 	"strings"
@@ -33,8 +32,6 @@ import (
 
 const paymentHeader = "Livepeer-Payment"
 const segmentHeader = "Livepeer-Segment"
-
-const pixelEstimateMultiplier = 1.02
 
 var errSegEncoding = errors.New("ErrorSegEncoding")
 var errSegSig = errors.New("ErrSegSig")
@@ -291,7 +288,7 @@ func SubmitSegment(sess *BroadcastSession, seg *stream.HLSSegment, nonce uint64)
 		return nil, err
 	}
 
-	fee, err := estimateFee(seg, sess.Profiles, priceInfo)
+	fee, err := core.EstimateFee(seg, sess.Profiles, priceInfo)
 	if err != nil {
 		return nil, err
 	}
@@ -505,39 +502,6 @@ func genSegCreds(sess *BroadcastSession, seg *stream.HLSSegment) (string, error)
 		return "", err
 	}
 	return base64.StdEncoding.EncodeToString(data), nil
-}
-
-func estimateFee(seg *stream.HLSSegment, profiles []ffmpeg.VideoProfile, priceInfo *big.Rat) (*big.Rat, error) {
-	if priceInfo == nil {
-		return nil, nil
-	}
-
-	// TODO: Estimate the number of input pixels
-	// Estimate the number of output pixels
-	var outPixels int64
-	for _, p := range profiles {
-		w, h, err := ffmpeg.VideoProfileResolution(p)
-		if err != nil {
-			return nil, err
-		}
-		framerate := p.Framerate
-		if framerate == 0 {
-			// FPS is being passed through (no fps adjustment)
-			// TODO incorporate the actual number of frames from the input
-			framerate = 120 // conservative estimate of input fps
-		}
-
-		// Take the ceiling of the duration to always overestimate
-		outPixels += int64(w*h) * int64(framerate) * int64(math.Ceil(seg.Duration))
-	}
-
-	// feeEstimate = pixels * pixelEstimateMultiplier * priceInfo
-	fee := new(big.Rat).SetInt64(outPixels)
-	// Multiply pixels by pixelEstimateMultiplier to ensure that we never underpay
-	fee.Mul(fee, new(big.Rat).SetFloat64(pixelEstimateMultiplier))
-	fee.Mul(fee, priceInfo)
-
-	return fee, nil
 }
 
 func newBalanceUpdate(sess *BroadcastSession, minCredit *big.Rat) (*BalanceUpdate, error) {
